phase(2.1.9.0): translocate AssignmentType enum to models/source.py (C9 STEP 0; K-stop-1 resolution)
Pre-C9 preparation per C9_INVESTIGATION.md Розділ K-stop-1: AssignmentType
enum (TEST/SHORT_TASK/TASK/PROJECT) lives in AuthoredDocument.task_type
ORM column + 6 imports in active document path. Translocate to
models/source.py alongside SourceType/MaterialRole/ChunkType (canonical
document taxonomy location). Methodist legacy layer is deleted in C9.3.

Files:
- src/course_supporter/models/source.py: +15 LOC (class AssignmentType
  with docstring noting C9.0 translocation lineage)
- src/course_supporter/models/methodist.py: -12 LOC (class removed +
  StrEnum import removed + Task Taxonomy section comment removed +
  re-import from source for surviving AssignmentRecommendation.assignment_type
  field type annotation)
- 5 import switches consolidated to single-line imports:
  storage/authored_document_repository.py:12,
  api/schemas.py:11,
  api/routes/documents.py:48,
  tests/unit/test_authored_document_repository.py:149, 188 (inline),
  tests/unit/test_methodist_schemas.py: AssignmentType split out of methodist import block

Verification (Rule #25):
- rg "from course_supporter.models.methodist import AssignmentType"
  src/ tests/ → 0 hits
- rg "from course_supporter.models.source import AssignmentType"
  src/ tests/ → 7 hits (5 src + 2 tests)
- AssignmentType class definition: only in src/course_supporter/models/source.py:26

Quality gates:
- ruff check src/ tests/: All checks passed
- ruff format --check: 7 files already formatted
- mypy --strict src/: clean (157 source files)
- pytest tests/ -q (Rule #22 default): 2357 passed, 200 skipped, 0 failed
